Feat/zod validation

+++ src/actions/actions.ts
@@ -6,44 +6,66 @@ import { revalidatePath } from "next/cache";
 import prisma from "@/lib/db";
 import { PetEssentials } from "@/lib/types";
 import { sleep } from "@/lib/utils";
+import { petFormSchema, petIdeaSchema } from "@/lib/validations";
 
-export async function addPet(newPetData: PetEssentials) {
+export async function addPet(newPetData: unknown) {
   await sleep(1000);
 
+  const validatedPet = petFormSchema.safeParse(newPetData);
+  if (!validatedPet.success) {
+    return { message: "Invalid data" };
+  }
   try {
     await prisma.pet.create({
-      data: newPetData,
+      data: validatedPet.data,
     });
   } catch (error) {
+    console.log(error);
+
     return { message: "error" };
   }
 
   revalidatePath("/app", "layout");
 }
 
-export async function editPet(petId: Pet["id"], updatedPetData: PetEssentials) {
+export async function editPet(petId: unknown, updatedPetData: unknown) {
   await sleep(1000);
+
+  const validatedPet = petFormSchema.safeParse(updatedPetData);
+  if (!validatedPet.success) {
+    return { message: "Invalid data" };
+  }
+
+  const validatedPetId = petIdeaSchema.safeParse(petId);
+  if (!validatedPetId.success) {
+    return { message: "Invalid data" };
+  }
 
   try {
     const updatedPet = await prisma.pet.update({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
-      data: updatedPetData,
+      data: validatedPet.data,
     });
   } catch (error) {
-    return { message: "Could not edit" };
+    return { message: "Could not edit pet" };
   }
   revalidatePath("/app", "layout");
 }
 
-export async function deletePet(petId: Pet["id"]) {
+export async function deletePet(petId: unknown) {
   await sleep(1000);
+
+  const validatedPetId = petIdeaSchema.safeParse(petId);
+  if (!validatedPetId.success) {
+    return { message: "Invalid data" };
+  }
 
   try {
     await prisma.pet.delete({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
     });
   } catch (error) {
